Cleaning up production dependencies by splitting dev and prod servers
