+++ package.json
@@ -32,12 +32,14 @@
     "body-parser": "^1.15.2",
     "chalk": "^1.1.3",
     "connect-memcached": "^0.2.0",
+    "css-modules-require-hook": "^4.0.5",
     "dotenv": "^2.0.0",
     "es6-promise": "^4.0.5",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
     "express-winston": "^2.0.0",
     "helmet": "^3.1.0",
+    "images-require-hook": "^1.0.3",
     "isomorphic-fetch": "^2.2.1",
     "nconf": "^0.8.4",
     "nconf-yaml": "^1.0.2",
@@ -73,7 +75,6 @@
     "babel-register": "^6.16.3",
     "chokidar": "^1.6.1",
     "css-loader": "^0.25.0",
-    "css-modules-require-hook": "^4.0.5",
     "dotenv": "^2.0.0",
     "enzyme": "^2.5.1",
     "eslint": "^3.8.1",
@@ -82,7 +83,6 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "flux-standard-action": "^1.0.0",
-    "images-require-hook": "^1.0.3",
     "mocha": "^3.1.2",
     "nock": "^9.0.2",
     "nyc": "^9.0.1",
