+++ src/server/index.js
@@ -1,14 +1,8 @@
-require('babel-register');
 require('css-modules-require-hook/preset');
-
-const logging = require('./logging/').default;
-const logger = logging.getLogger('express');
-
-const conf = require('./helpers/config').conf;
-const WEBPACK_DEV_URL = conf.get('SERVER:WEBPACK_DEV_URL') || '';
-require('images-require-hook')('.svg', `${WEBPACK_DEV_URL}/static/icons`);
+require('images-require-hook')('.svg', '/static/icons');
 
 const app = require('./server').default;
+const logger = app.get('logger');
 
 const server = app.listen(app.locals.port, app.locals.host, () => {
   const host = server.address().address;
